DataLoader/requirements.txt
greenlet==1.0.0
grpcio==1.36.1
grpcio-tools==1.36.1
importlib-metadata==3.7.3
install==1.3.4
pip==21.0.1
protobuf==3.15.6
psycopg2==2.8.6
pyaes==1.6.1
pyasn1==0.4.8
rsa==4.7.2
setuptools==47.1.0
six==1.15.0
SQLAlchemy==1.4.1
Telethon==1.20
telethon-session-sqlalchemy==0.2.16
typing-extensions==3.7.4.3
zipp==3.4.1
